+++ src/main/java/com/commercetools/sync/categories/utils/CategoryReferenceReplacementUtils.java
@@ -6,6 +6,7 @@ import io.sphere.sdk.categories.CategoryDraftBuilder;
 import io.sphere.sdk.categories.expansion.CategoryExpansionModel;
 import io.sphere.sdk.categories.queries.CategoryQuery;
 import io.sphere.sdk.expansion.ExpansionPath;
+import io.sphere.sdk.models.AssetDraft;
 import io.sphere.sdk.models.ResourceIdentifier;
 import io.sphere.sdk.queries.QueryExecutionUtils;
 import io.sphere.sdk.types.CustomFieldsDraft;
@@ -14,17 +15,18 @@ import javax.annotation.Nonnull;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static com.commercetools.sync.commons.utils.SyncUtils.replaceAssetsReferencesIdsWithKeys;
 import static com.commercetools.sync.commons.utils.SyncUtils.replaceCustomTypeIdWithKeys;
 import static com.commercetools.sync.commons.utils.SyncUtils.replaceReferenceIdWithKey;
 
 public final class CategoryReferenceReplacementUtils {
 
     /**
-     * Takes a list of Categories that are supposed to have their custom type and parent category reference expanded
-     * in order to be able to fetch the keys and replace the reference ids with the corresponding keys and then return
-     * a new list of category drafts with their references containing keys instead of the ids. Note that if the
-     * references are not expanded for a category, the reference ids will not be replaced with keys and will still have
-     * their ids in place.
+     * Takes a list of Categories that are supposed to have their custom type, parent category and asset custom type
+     * references expanded in order to be able to fetch the keys and replace the reference ids with the corresponding
+     * keys and then return a new list of category drafts with their references containing keys instead of the ids.
+     * Note that if the references are not expanded for a category, the reference ids will not be replaced with keys and
+     * will still have their ids in place.
      *
      * @param categories the categories to replace their reference ids with keys
      * @return a list of category drafts with keys instead of ids for references.
@@ -38,10 +40,13 @@ public final class CategoryReferenceReplacementUtils {
                 @SuppressWarnings("ConstantConditions") // NPE checked in replaceReferenceIdWithKey
                 final ResourceIdentifier<Category> parentWithKeyInReference = replaceReferenceIdWithKey(
                     category.getParent(), () -> Category.referenceOfId(category.getParent().getObj().getKey()));
+                final List<AssetDraft> assetDraftsWithKeyInReference =
+                    replaceAssetsReferencesIdsWithKeys(category.getAssets());
 
                 return CategoryDraftBuilder.of(category)
                                            .custom(customTypeWithKeysInReference)
                                            .parent(parentWithKeyInReference)
+                                           .assets(assetDraftsWithKeyInReference)
                                            .build();
             })
             .collect(Collectors.toList());
@@ -52,6 +57,7 @@ public final class CategoryReferenceReplacementUtils {
      * expanded for the sync:
      * <ul>
      *     <li>Custom Type</li>
+     *     <li>Assets Custom Types</li>
      *     <li>Parent Category</li>
      * </ul>
      *
@@ -62,6 +68,7 @@ public final class CategoryReferenceReplacementUtils {
         return CategoryQuery.of()
                             .withLimit(QueryExecutionUtils.DEFAULT_PAGE_SIZE)
                             .withExpansionPaths(ExpansionPath.of("custom.type"))
+                            .plusExpansionPaths(ExpansionPath.of("assets[*].custom.type"))
                             .plusExpansionPaths(CategoryExpansionModel::parent);
     }